+++ params/params.go
@@ -97,6 +97,7 @@ func ParseParams() (p *Params, err error) {
 	return p, nil
 }
 
+// Проверяет корректность ввода номера института
 func checkDep(p *Params) error {
 	if p.Dep < 1 || p.Dep > 14 || p.Dep == 13 {
 		return ErrDepOutOfBound
@@ -104,6 +105,7 @@ func checkDep(p *Params) error {
 	return nil
 }
 
+// Проверяет корректность ввода номера курса
 func checkCourse(p *Params) error {
 	if p.Course < 1 || p.Course > 6 {
 		return ErrCourseOutOfBound
@@ -112,6 +114,7 @@ func checkCourse(p *Params) error {
 	return nil
 }
 
+// Проверяет корректность ввода номера недели
 func checkWeek(p *Params) error {
 	if p.Session {
 		p.Week = 0
