+++ params/strings.go
@@ -1,7 +1,6 @@
 package params
 
 // Строки для справки
-
 const (
 	versionDesc = "Печать номера версии и выход"
 	versionText = "timetable 1.0.3b2\n" +
